feat: swap out components for single page create design
