+++ src/shared/containers/index.tsx
@@ -103,10 +103,6 @@ export const CreateSubscriptionForm = lazy(() =>
   import('src/writeData/subscriptions/components/CreateSubscriptionPage')
 )
 
-export const SubscriptionDetailsPage = lazy(() =>
-  import('src/writeData/subscriptions/components/SubscriptionDetailsPage')
-)
-
 export const HomepageContainer = lazy(() =>
   import(
     'src/homepageExperience/containers/HomepageContainer'
