+++ src/writeData/subscriptions/components/BrokerDetails.scss
@@ -2,6 +2,26 @@
 .update-broker-form {
   float: right;
   width: 75%;
+
+  &__fixed {
+    position: fixed;
+    background-color: $cf-grey-5;
+    background-attachment: fixed;
+    width: 61%;
+    z-index: 9999;
+    padding-bottom: 20px;
+
+    &__broker-buttons {
+      padding-top: 46px;
+      padding-left: 24px;
+      padding-right: 24px;
+
+      button {
+        margin-left: 8px;
+      }
+    }
+  }
+
   .cf-overlay--body {
     padding-top: 50px;
   }
@@ -27,21 +47,6 @@
   input[type='number']::-webkit-outer-spin-button {
     -webkit-appearance: none;
     margin: 0;
-  }
-  &__premium-container {
-    margin-left: $cf-space-s;
-    .cf-icon {
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
-    &__text {
-      margin-left: $cf-space-s;
-      text-transform: uppercase;
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
   }
   &__header {
     text-transform: uppercase;
@@ -83,24 +88,5 @@
     margin-bottom: 60px;
     margin-left: 12px;
     margin-right: 12px;
-  }
-
-  &__fixed {
-    position: fixed;
-    background-color: $cf-grey-5;
-    background-attachment: fixed;
-    width: 61%;
-    z-index: 9999;
-    padding-bottom: 20px;
-
-    &__broker-buttons {
-      padding-top: 46px;
-      padding-left: 24px;
-      padding-right: 24px;
-
-      button {
-        margin-left: 8px;
-      }
-    }
   }
 }
