+++ src/writeData/subscriptions/components/BrokerForm.scss
@@ -2,6 +2,26 @@
 .create-broker-form {
   float: right;
   width: 75%;
+
+  &__fixed {
+    position: fixed;
+    background-color: $cf-grey-5;
+    background-attachment: fixed;
+    width: 61%;
+    z-index: 9999;
+    padding-bottom: 20px;
+
+    &__broker-buttons {
+      padding-top: 46px;
+      padding-left: 24px;
+      padding-right: 24px;
+
+      button {
+        margin-left: 8px;
+      }
+    }
+  }
+
   .cf-flex-box__margin-md.cf-flex-box__row.cf-flex-box__justify-flex-start > * {
     margin-right: 0;
   }
@@ -37,10 +57,31 @@
       -webkit-text-fill-color: transparent;
     }
   }
+  &__upgrade-button {
+    background: linear-gradient(45deg, #0098f0 0%, $c-pulsar 100%) !important;
+    color: $cf-white;
+    padding: 0 $cf-space-xs;
+    height: 40px;
+    font-weight: 500;
+    font-family: 'Proxima Nova', Helvetica, Arial, Tahoma, Verdana, sans-serif;
+    border-style: solid;
+    border-width: 0;
+    text-decoration: none;
+    transition: background-color 200ms cubic-bezier(0.18, 0.16, 0.2, 1),
+      box-shadow 200ms cubic-bezier(0.18, 0.16, 0.2, 1),
+      color 200ms cubic-bezier(0.18, 0.16, 0.2, 1);
+    outline: none;
+    border-radius: 2px;
+    white-space: nowrap;
+    position: relative;
+    z-index: 1;
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+  }
   &__header {
     text-transform: uppercase;
-    margin-top: $cf-space-2xl;
-    margin-bottom: $cf-space-l;
+    margin: $cf-space-l 0;
   }
   &__text {
     color: $cf-grey-95;
@@ -65,5 +106,15 @@
         margin-bottom: 0;
       }
     }
+  }
+  &__line {
+    background: linear-gradient(
+      256.11deg,
+      rgba(147, 148, 255, 0.2) 0%,
+      rgba(81, 60, 198, 0.2) 100%
+    );
+    height: 3px;
+    margin: 0 12px;
+    margin-bottom: 30px;
   }
 }
