+++ src/writeData/subscriptions/components/ParsingDetails.tsx
@@ -1,6 +1,5 @@
 // Libraries
 import React, {FC} from 'react'
-import {useSelector} from 'react-redux'
 
 // Components
 import {Grid, Form, Overlay} from '@influxdata/clockface'
@@ -9,11 +8,6 @@ import StringParsingForm from 'src/writeData/subscriptions/components/StringPars
 import JsonParsingForm from 'src/writeData/subscriptions/components/JsonParsingForm'
 import ParsingDetailsEdit from 'src/writeData/subscriptions/components/ParsingDetailsEdit'
 import ParsingDetailsReadOnly from 'src/writeData/subscriptions/components/ParsingDetailsReadOnly'
-import StatusHeader from 'src/writeData/subscriptions/components/StatusHeader'
-import DetailsFormFooter from 'src/writeData/subscriptions/components/DetailsFormFooter'
-
-// Utils
-import {getOrg} from 'src/organizations/selectors'
 
 // Types
 import {Subscription} from 'src/types/subscriptions'
@@ -24,87 +18,49 @@ import 'src/writeData/subscriptions/components/ParsingDetails.scss'
 interface Props {
   currentSubscription: Subscription
   updateForm: (any) => void
-  saveForm: (any) => void
   edit: boolean
-  setEdit: (any) => void
-  singlePage: boolean
-  setStatus: (any) => void
-  setFormActive: (any) => void
-  active: string
 }
 
-const ParsingDetails: FC<Props> = ({
-  currentSubscription,
-  updateForm,
-  saveForm,
-  edit,
-  setEdit,
-  singlePage,
-  setStatus,
-  setFormActive,
-  active,
-}) => {
-  const org = useSelector(getOrg)
-  return (
-    <div className="update-parsing-form" id="parsing">
-      <Form onSubmit={() => {}} testID="update-parsing-form-overlay">
-        {!singlePage && (
-          <StatusHeader
-            currentSubscription={currentSubscription}
-            setStatus={setStatus}
-          />
-        )}
-        <Overlay.Header title="Define Data Parsing Rules"></Overlay.Header>
-        <Overlay.Body>
-          <Grid>
-            <Grid.Row>
-              {edit ? (
-                <ParsingDetailsEdit
-                  currentSubscription={currentSubscription}
-                  updateForm={updateForm}
-                  className="update"
-                />
-              ) : (
-                <ParsingDetailsReadOnly
-                  currentSubscription={currentSubscription}
-                />
-              )}
-              {currentSubscription.dataFormat === 'string' && (
-                <StringParsingForm
-                  formContent={currentSubscription}
-                  updateForm={updateForm}
-                  edit={edit}
-                />
-              )}
-              {currentSubscription.dataFormat === 'json' && (
-                <JsonParsingForm
-                  formContent={currentSubscription}
-                  updateForm={updateForm}
-                  edit={edit}
-                />
-              )}
-              {currentSubscription.dataFormat === 'lineprotocol' && (
-                <LineProtocolForm />
-              )}
-            </Grid.Row>
-          </Grid>
-        </Overlay.Body>
-        {!singlePage ? (
-          <DetailsFormFooter
-            nextForm=""
-            id={org.id}
-            edit={edit}
-            setEdit={setEdit}
-            setFormActive={setFormActive}
-            formActive={active}
-            currentSubscription={currentSubscription}
-            saveForm={saveForm}
-          />
-        ) : (
-          <div className="update-parsing-form__line"></div>
-        )}
-      </Form>
-    </div>
-  )
-}
+const ParsingDetails: FC<Props> = ({currentSubscription, updateForm, edit}) => (
+  <div className={currentSubscription.dataFormat === 'lineprotocol' ? "update-parsing-form--line-protocol" : "update-parsing-form"} id="parsing">
+    <Form onSubmit={() => {}} testID="update-parsing-form-overlay">
+      <Overlay.Header title="Define Data Parsing Rules"></Overlay.Header>
+      <Overlay.Body>
+        <Grid>
+          <Grid.Row>
+            {edit ? (
+              <ParsingDetailsEdit
+                currentSubscription={currentSubscription}
+                updateForm={updateForm}
+                className="update"
+              />
+            ) : (
+              <ParsingDetailsReadOnly
+                currentSubscription={currentSubscription}
+              />
+            )}
+            {currentSubscription.dataFormat === 'string' && (
+              <StringParsingForm
+                formContent={currentSubscription}
+                updateForm={updateForm}
+                edit={edit}
+              />
+            )}
+            {currentSubscription.dataFormat === 'json' && (
+              <JsonParsingForm
+                formContent={currentSubscription}
+                updateForm={updateForm}
+                edit={edit}
+              />
+            )}
+            {currentSubscription.dataFormat === 'lineprotocol' && (
+              <LineProtocolForm />
+            )}
+          </Grid.Row>
+        </Grid>
+      </Overlay.Body>
+      <div className="update-parsing-form__line"></div>
+    </Form>
+  </div>
+)
 export default ParsingDetails
