+++ src/writeData/subscriptions/components/ParsingForm.scss
@@ -2,55 +2,36 @@
 .create-parsing-form {
   float: right;
   width: 75%;
-  .cf-overlay--header {
-    justify-content: unset;
-  }
   .cf-form--wrapper {
-    padding-bottom: $cf-space-xl;
+    padding-bottom: 50px;
   }
+
+  &--line-protocol {
+    float: right;
+    width: 75%;
+    .cf-form--wrapper {
+    padding-bottom: 550px;
+    }
+}
+  .cf-overlay--header {
+      justify-content: unset;
+    }
+ 
   &__text {
     color: $cf-grey-95;
   }
   &__header {
     text-transform: uppercase;
-    margin-top: $cf-space-2xl;
-    margin-bottom: $cf-space-l;
+    margin: $cf-space-l 0;
   }
-  &__premium-container {
-    margin-left: $cf-space-s;
-    .cf-icon {
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
-    &__text {
-      margin-left: $cf-space-s;
-      text-transform: uppercase;
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
-  }
-  &__upgrade-button {
-    background: linear-gradient(45deg, #0098f0 0%, $c-pulsar 100%) !important;
-    color: $cf-white;
-    padding: 0 $cf-space-xs;
-    height: 40px;
-    font-weight: 500;
-    font-family: 'Proxima Nova', Helvetica, Arial, Tahoma, Verdana, sans-serif;
-    border-style: solid;
-    border-width: 0;
-    text-decoration: none;
-    transition: background-color 200ms cubic-bezier(0.18, 0.16, 0.2, 1),
-      box-shadow 200ms cubic-bezier(0.18, 0.16, 0.2, 1),
-      color 200ms cubic-bezier(0.18, 0.16, 0.2, 1);
-    outline: none;
-    border-radius: 2px;
-    white-space: nowrap;
-    position: relative;
-    z-index: 1;
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
+  &__line {
+    background: linear-gradient(
+      256.11deg,
+      rgba(147, 148, 255, 0.2) 0%,
+      rgba(81, 60, 198, 0.2) 100%
+    );
+    height: 3px;
+    margin: 0 12px;
+    margin-bottom: 30px;
   }
 }
