+++ src/writeData/subscriptions/components/SubscriptionForm.scss
@@ -16,28 +16,22 @@
   .cf-flex-box__margin-md.cf-flex-box__row.cf-flex-box__justify-flex-start > * {
     margin-right: 0;
   }
-  &__premium-container {
-    margin-left: $cf-space-s;
-    .cf-icon {
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
-    &__text {
-      margin-left: $cf-space-s;
-      text-transform: uppercase;
-      background: -webkit-linear-gradient(45deg, #00a3ff 0%, #34bb55 100%);
-      -webkit-background-clip: text;
-      -webkit-text-fill-color: transparent;
-    }
-  }
   &__text {
     color: $cf-grey-95;
     margin-bottom: $cf-space-l;
   }
   &__header {
     text-transform: uppercase;
-    margin-top: $cf-space-2xl;
-    margin-bottom: $cf-space-l;
+    margin: $cf-space-l 0;
+  }
+  &__line {
+    background: linear-gradient(
+      256.11deg,
+      rgba(147, 148, 255, 0.2) 0%,
+      rgba(81, 60, 198, 0.2) 100%
+    );
+    height: 3px;
+    margin: 0 12px;
+    margin-bottom: 30px;
   }
 }
